add fuel js and php was moved to external files from fuel.php
